+++ package.json
@@ -16,6 +16,7 @@
     "classnames": "^2.3.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^6.2.1",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
