+++ src/MatchUnit.tsx
@@ -10,50 +10,19 @@ import { v4 as uuidv4 } from 'uuid';
 type MatchUnitProps = {
   animation: CSSProperties;
   userId: string;
+  matches: Array<any>;
 };
 
-export default function MatchUnit({ animation, userId }: MatchUnitProps) {
+export default function MatchUnit({ animation, userId, matches }: MatchUnitProps) {
   const [mode, setMode] = useState<string>('');
-  const [matches, setMatches] = useState<Array<any>>([]);
 
   useEffect(() => {
     setMode('unit');
-    if (userId !== '') {
-      getMatch();
-      const matchSubscriptionUserA = supabase
-        .from(`match:userA=eq.${userId}`)
-        .on('*', payload => {
-          console.log('changed', payload);
-        })
-        .subscribe();
-
-      const matchSubscriptionUserB = supabase
-        .from(`match:userB=eq.${userId}`)
-        .on('*', payload => {
-          console.log('changed', payload);
-        })
-        .subscribe();
-      return () => {
-        supabase.removeSubscription(matchSubscriptionUserA);
-        supabase.removeSubscription(matchSubscriptionUserB);
-      };
-    }
   }, []);
 
   useEffect(() => {
     console.log(matches);
   }, [matches]);
-
-  const getMatch = async () => {
-    try {
-      const { data, error } = await supabase.rpc('f_getmatchinfo', {
-        selfid: userId,
-      });
-      setMatches(data as Array<any>);
-    } catch (error) {
-      console.log(error);
-    }
-  };
 
   return (
     <>
